Rename recording kind to genre
