+++ .docker_files/main/__manifest__.py
@@ -17,9 +17,9 @@
         'recording',
         'recording_artist',
         'recording_contributor',
-        'recording_kind',
+        'recording_genre',
         'recording_lang',
-        'recording_subkind',
+        'recording_subgenre',
         'recording_tag',
         'recording_url_link',
         'recording_version',
