+++ recording_genre/views/menu.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<odoo>
+
+    <menuitem
+        id="menu_config_recording_genre"
+        name="Genres"
+        parent="recording.menu_config_recording"
+        action="action_recording_genre"
+        sequence="50"
+        />
+
+</odoo>
