+++ recording_genre/views/recording.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<odoo>
+
+    <record id="recording_form_with_genre" model="ir.ui.view">
+        <field name="name">Recording Form: Add record genre</field>
+        <field name="model">recording</field>
+        <field name="inherit_id" ref="recording.recording_form"/>
+        <field name="arch" type="xml">
+            <div name="right_column" position="inside">
+                <group name="genre" string="Genre">
+                    <field name="genre_id"/>
+                </group>
+            </div>
+        </field>
+    </record>
+
+</odoo>
