--- recording_kind/README.rst
@@ -1,26 +1,0 @@
-Recording Kinds
-===============
-
-This module adds recording kinds.
-
-Usage
------
-
-The configuration kinds menu item is available to members of the group ``Recording / Manager``
-through the ``Recording / Configuration`` menu:
-
-It shows the list of kinds.
-
-.. image:: static/description/kinds.png
-
-By clicking on a create button you can create a new kind of recording.
-
-Recording
----------
-On the form view of a recording, a field is added to select the kind of recording.
-
-.. image:: static/description/recording_form.png
-
-Contributors
-------------
-* Numigi (tm) and all its contributors (https://bit.ly/numigiens)
