--- recording_kind/i18n/fr.po
@@ -1,112 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* recording_kind
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: \n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2019-12-18 20:57+0000\n"
-"PO-Revision-Date: 2019-12-18 15:58-0500\n"
-"Last-Translator: <>\n"
-"Language-Team: \n"
-"Language: fr\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Project-Id-Type: Odoo Server 12.0\n"
-"MIME-Type: 1.0\n"
-"Plural-Forms: \n"
-"X-Generator: Poedit 2.0.6\n"
-
-#. module: recording_kind
-#: model:ir.model.fields,field_description:recording_kind.field_recording_kind__active
-msgid "Active"
-msgstr "Actif"
-
-#. module: recording_kind
-#: model_terms:ir.ui.view,arch_db:recording_kind.recording_kind_search
-msgid "Archived"
-msgstr "Archivé"
-
-#. module: recording_kind
-#: model:ir.model.fields,field_description:recording_kind.field_recording_kind__create_uid
-msgid "Created by"
-msgstr "Crée par"
-
-#. module: recording_kind
-#: model:ir.model.fields,field_description:recording_kind.field_recording_kind__create_date
-msgid "Created on"
-msgstr "Crée le"
-
-#. module: recording_kind
-#: model:ir.model.fields,field_description:recording_kind.field_recording_kind__display_name
-msgid "Display Name"
-msgstr "Nom"
-
-#. module: recording_kind
-#: model:ir.model.fields,field_description:recording_kind.field_recording_kind__id
-msgid "ID"
-msgstr "ID"
-
-#. module: recording_kind
-#: model:ir.model.fields,field_description:recording_kind.field_recording__kind_id
-#: model_terms:ir.ui.view,arch_db:recording_kind.recording_form_with_kind
-msgid "Kind"
-msgstr "Genre"
-
-#. module: recording_kind
-#: model:ir.model,name:recording_kind.model_recording_kind
-msgid "Kind of Recording"
-msgstr "Genre d'enregistrement"
-
-#. module: recording_kind
-#: model:ir.ui.menu,name:recording_kind.menu_config_recording_kind
-msgid "Kinds"
-msgstr "Genres"
-
-#. module: recording_kind
-#: model:ir.model.fields,field_description:recording_kind.field_recording_kind____last_update
-msgid "Last Modified on"
-msgstr "Dernière modification le"
-
-#. module: recording_kind
-#: model:ir.model.fields,field_description:recording_kind.field_recording_kind__write_uid
-msgid "Last Updated by"
-msgstr "Dernière mise à jour par"
-
-#. module: recording_kind
-#: model:ir.model.fields,field_description:recording_kind.field_recording_kind__write_date
-msgid "Last Updated on"
-msgstr "Dernière mise à jour le"
-
-#. module: recording_kind
-#: model:ir.model.fields,field_description:recording_kind.field_recording_kind__name
-msgid "Name"
-msgstr "Nom"
-
-#. module: recording_kind
-#: model_terms:ir.ui.view,arch_db:recording_kind.recording_kind_form
-msgid "Record Kind"
-msgstr "Genre d'enregistrement"
-
-#. module: recording_kind
-#: model:ir.actions.act_window,name:recording_kind.action_recording_kind
-#: model_terms:ir.ui.view,arch_db:recording_kind.recording_kind_list
-msgid "Record Kinds"
-msgstr "Genres d'enregistrements"
-
-#. module: recording_kind
-#: model:ir.model,name:recording_kind.model_recording
-msgid "Recording"
-msgstr "Enregistrement"
-
-#. module: recording_kind
-#: model_terms:ir.ui.view,arch_db:recording_kind.recording_kind_search
-msgid "Search Record Kinds"
-msgstr "Chercher des genres d'enregistrements"
-
-#. module: recording_kind
-#: model:ir.model.fields,field_description:recording_kind.field_recording_kind__sequence
-msgid "Sequence"
-msgstr "Séquence"
