--- recording_kind/views/menu.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<odoo>
-
-    <menuitem
-        id="menu_config_recording_kind"
-        name="Kinds"
-        parent="recording.menu_config_recording"
-        action="action_recording_kind"
-        sequence="50"
-        />
-
-</odoo>
