--- recording_kind/views/recording.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<odoo>
-
-    <record id="recording_form_with_kind" model="ir.ui.view">
-        <field name="name">Recording Form: Add record kind</field>
-        <field name="model">recording</field>
-        <field name="inherit_id" ref="recording.recording_form"/>
-        <field name="arch" type="xml">
-            <div name="right_column" position="inside">
-                <group name="kind" string="Kind">
-                    <field name="kind_id"/>
-                </group>
-            </div>
-        </field>
-    </record>
-
-</odoo>
