+++ recording_subgenre/models/recording_subgenre.py
@@ -1,0 +1,15 @@
+# © 2019 - today Numigi (tm) and all its contributors (https://bit.ly/numigiens)
+# License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
+
+from odoo import fields, models
+
+
+class RecordSubGenre(models.Model):
+
+    _name = 'recording.subgenre'
+    _description = 'Subgenre of Recording'
+    _order = 'sequence'
+
+    sequence = fields.Integer()
+    name = fields.Char(string='Name', required=True, translate=True)
+    active = fields.Boolean(string='Active', default=True)
