+++ recording_subgenre/security/ir.model.access.csv
@@ -1,0 +1,3 @@
+id,name,model_id/id,group_id/id,perm_read,perm_write,perm_create,perm_unlink
+access_recording_subgenre_user,name_recording_subgenre_user,model_recording_subgenre,recording.group_user,1,0,0,0
+access_recording_subgenre_manager,name_recording_subgenre_manager,model_recording_subgenre,recording.group_manager,1,1,1,0
