+++ recording_subgenre/views/recording.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<odoo>
+
+    <record id="recording_form_with_subgenre" model="ir.ui.view">
+        <field name="name">Recording Form: Add record subgenre</field>
+        <field name="model">recording</field>
+        <field name="inherit_id" ref="recording_genre.recording_form_with_genre"/>
+        <field name="arch" type="xml">
+            <field name="genre_id" position="after">
+                <field name="subgenre_id"/>
+            </field>
+        </field>
+    </record>
+
+</odoo>
