+++ recording_subgenre/views/recording_subgenre.xml
@@ -1,0 +1,72 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<odoo>
+
+    <record id="recording_subgenre_form" model="ir.ui.view">
+        <field name="name">Record Subgenre Form</field>
+        <field name="model">recording.subgenre</field>
+        <field name="arch" type="xml">
+            <form string="Record Subgenre">
+                <header>
+                </header>
+                <sheet>
+                    <div class="oe_button_box" name="button_box">
+                        <button name="toggle_active" type="object"
+                                class="oe_stat_button" icon="fa-archive">
+                            <field name="active" widget="boolean_button"
+                                options='{"terminology": "archive"}'/>
+                        </button>
+                    </div>
+                    <div class="oe_title">
+                        <h1>
+                            <field name="name"/>
+                        </h1>
+                    </div>
+                    <group>
+                        <group name="group_1">
+                            <field name="sequence"/>
+                        </group>
+                        <group name="group_2">
+                        </group>
+                    </group>
+                    <notebook>
+                    </notebook>
+                </sheet>
+            </form>
+        </field>
+    </record>
+
+    <record id="recording_subgenre_list" model="ir.ui.view">
+        <field name="name">Record Subgenre List</field>
+        <field name="model">recording.subgenre</field>
+        <field name="arch" type="xml">
+            <tree string="Record Subgenres" editable="top">
+                <field name="sequence" widget="handle"/>
+                <field name="name"/>
+                <field name="active"/>
+            </tree>
+        </field>
+    </record>
+
+    <record id="recording_subgenre_search" model="ir.ui.view">
+        <field name="name">Record Subgenre Search</field>
+        <field name="model">recording.subgenre</field>
+        <field name="arch" type="xml">
+            <search string="Search Record Subgenres">
+                <field name="name"/>
+                <separator/>
+                <filter string="Archived" name="inactive" domain="[('active', '=', False)]"/>
+           </search>
+        </field>
+    </record>
+
+    <record id="action_recording_subgenre" model="ir.actions.act_window">
+        <field name="name">Record Subgenres</field>
+        <field name="type">ir.actions.act_window</field>
+        <field name="res_model">recording.subgenre</field>
+        <field name="view_type">form</field>
+        <field name="view_mode">tree,form</field>
+        <field name="view_id" ref="recording_subgenre_list"/>
+        <field name="search_view_id" ref="recording_subgenre_search"/>
+    </record>
+
+</odoo>
