--- recording_subkind/README.rst
@@ -1,26 +1,0 @@
-Recording Subkinds
-==================
-
-This module adds recording subkinds.
-
-Usage
------
-
-The configuration subkinds menu item is available to members of the group ``Recording / Manager``
-through the ``Recording / Configuration`` menu:
-
-It shows the list of subkinds.
-
-.. image:: static/description/subkinds.png
-
-By clicking on a create button you can create a new subkind of recording.
-
-Recording
----------
-On the form view of a recording, a field is added to select the subkind of recording.
-
-.. image:: static/description/recording_form.png
-
-Contributors
-------------
-* Numigi (tm) and all its contributors (https://bit.ly/numigiens)
