--- recording_subkind/i18n/fr.po
@@ -1,111 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* recording_subkind
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: \n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2019-12-18 20:58+0000\n"
-"PO-Revision-Date: 2019-12-18 15:59-0500\n"
-"Last-Translator: <>\n"
-"Language-Team: \n"
-"Language: fr\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Project-Id-Type: Odoo Server 12.0\n"
-"MIME-Type: 1.0\n"
-"Plural-Forms: \n"
-"X-Generator: Poedit 2.0.6\n"
-
-#. module: recording_subkind
-#: model:ir.model.fields,field_description:recording_subkind.field_recording_subkind__active
-msgid "Active"
-msgstr "Actif"
-
-#. module: recording_subkind
-#: model_terms:ir.ui.view,arch_db:recording_subkind.recording_subkind_search
-msgid "Archived"
-msgstr "Archivé"
-
-#. module: recording_subkind
-#: model:ir.model.fields,field_description:recording_subkind.field_recording_subkind__create_uid
-msgid "Created by"
-msgstr "Crée par"
-
-#. module: recording_subkind
-#: model:ir.model.fields,field_description:recording_subkind.field_recording_subkind__create_date
-msgid "Created on"
-msgstr "Crée le"
-
-#. module: recording_subkind
-#: model:ir.model.fields,field_description:recording_subkind.field_recording_subkind__display_name
-msgid "Display Name"
-msgstr "Nom"
-
-#. module: recording_subkind
-#: model:ir.model.fields,field_description:recording_subkind.field_recording_subkind__id
-msgid "ID"
-msgstr "ID"
-
-#. module: recording_subkind
-#: model:ir.model.fields,field_description:recording_subkind.field_recording_subkind____last_update
-msgid "Last Modified on"
-msgstr "Dernière modification le"
-
-#. module: recording_subkind
-#: model:ir.model.fields,field_description:recording_subkind.field_recording_subkind__write_uid
-msgid "Last Updated by"
-msgstr "Dernière mise à jour par"
-
-#. module: recording_subkind
-#: model:ir.model.fields,field_description:recording_subkind.field_recording_subkind__write_date
-msgid "Last Updated on"
-msgstr "Dernière mise à jour le"
-
-#. module: recording_subkind
-#: model:ir.model.fields,field_description:recording_subkind.field_recording_subkind__name
-msgid "Name"
-msgstr "Nom"
-
-#. module: recording_subkind
-#: model_terms:ir.ui.view,arch_db:recording_subkind.recording_subkind_form
-msgid "Record Subkind"
-msgstr "Sous-genre d'enregistrement"
-
-#. module: recording_subkind
-#: model:ir.actions.act_window,name:recording_subkind.action_recording_subkind
-#: model_terms:ir.ui.view,arch_db:recording_subkind.recording_subkind_list
-msgid "Record Subkinds"
-msgstr "Sous-genres d'enregistrements"
-
-#. module: recording_subkind
-#: model:ir.model,name:recording_subkind.model_recording
-msgid "Recording"
-msgstr "Enregistrement"
-
-#. module: recording_subkind
-#: model_terms:ir.ui.view,arch_db:recording_subkind.recording_subkind_search
-msgid "Search Record Subkinds"
-msgstr "Chercher des sous-genres d'enregistrements"
-
-#. module: recording_subkind
-#: model:ir.model.fields,field_description:recording_subkind.field_recording_subkind__sequence
-msgid "Sequence"
-msgstr "Séquence"
-
-#. module: recording_subkind
-#: model:ir.model.fields,field_description:recording_subkind.field_recording__subkind_id
-msgid "Subkind"
-msgstr "Sous-genre"
-
-#. module: recording_subkind
-#: model:ir.model,name:recording_subkind.model_recording_subkind
-msgid "Subkind of Recording"
-msgstr "Sous-genre d'enregistrement"
-
-#. module: recording_subkind
-#: model:ir.ui.menu,name:recording_subkind.menu_config_recording_subkind
-msgid "Subkinds"
-msgstr "Sous-genres"
