--- recording_subkind/models/__init__.py
@@ -1,7 +1,0 @@
-# © 2019 - today Numigi (tm) and all its contributors (https://bit.ly/numigiens)
-# License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
-
-from . import (
-    recording_subkind,
-    recording,
-)
