--- recording_subkind/views/menu.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<odoo>
-
-    <menuitem
-        id="menu_config_recording_subkind"
-        name="Subkinds"
-        parent="recording.menu_config_recording"
-        action="action_recording_subkind"
-        sequence="51"
-        />
-
-</odoo>
