--- recording_subkind/views/recording.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<odoo>
-
-    <record id="recording_form_with_subkind" model="ir.ui.view">
-        <field name="name">Recording Form: Add record subkind</field>
-        <field name="model">recording</field>
-        <field name="inherit_id" ref="recording_kind.recording_form_with_kind"/>
-        <field name="arch" type="xml">
-            <field name="kind_id" position="after">
-                <field name="subkind_id"/>
-            </field>
-        </field>
-    </record>
-
-</odoo>
